fix(admin): address review findings on bulk extend trial
- Add csvField() RFC 4180 escaping helper to admin-csv.ts and apply it
  in all three export functions in KiloclawExtendTrial so error messages
  with commas/quotes can't corrupt downloaded CSVs
- Parallelize DB updates in extendTrials via Promise.allSettled, extracting
  per-user logic into processOneEmail; eliminates up to 1000 sequential
  round-trips for large batches
- Align at_limit ceiling math with SQL interval '1 year': use calendar-year
  setFullYear arithmetic instead of fixed 365-day ms offset
- Compare at UTC-day granularity in the beyondCeiling check so ms-level
  clock drift cannot let an exact-ceiling row slip through as eligible
- Extract parseCsvToTable, extractEmailsFromColumn, guessEmailColumn, and
  parseEmailList from KiloclawExtendTrial into admin-csv.ts; remove the
  redundant CsvData type alias (use CsvTableData directly)
