+++ src/app/admin/components/KiloclawExtendTrial.tsx
@@ -36,14 +36,17 @@ import {
   RotateCcw,
   Download,
 } from 'lucide-react';
-import { downloadCsv } from '@/lib/admin-csv';
+import {
+  downloadCsv,
+  csvField,
+  parseCsvToTable,
+  extractEmailsFromColumn,
+  guessEmailColumn,
+  parseEmailList,
+  type CsvTableData,
+} from '@/lib/admin-csv';
 
 // --- Types ---
-
-type CsvData = {
-  headers: string[];
-  rows: Record<string, string>[];
-};
 
 type TrialResult = {
   email: string;
@@ -56,113 +59,6 @@ type TrialResult = {
 };
 
 type InputMode = 'paste' | 'csv';
-
-// --- CSV Parsing ---
-
-function parseCsvToTable(text: string): CsvData {
-  const lines = text.trim().split('\n');
-  if (lines.length === 0) return { headers: [], rows: [] };
-
-  const parseLine = (line: string): string[] => {
-    const parts: string[] = [];
-    let current = '';
-    let inQuotes = false;
-    for (let i = 0; i < line.length; i++) {
-      const ch = line[i];
-      if (inQuotes) {
-        if (ch === '"') {
-          if (line[i + 1] === '"') {
-            current += '"';
-            i++;
-          } else {
-            inQuotes = false;
-          }
-        } else {
-          current += ch;
-        }
-      } else {
-        if (ch === '"') {
-          inQuotes = true;
-        } else if (ch === ',') {
-          parts.push(current.trim());
-          current = '';
-        } else {
-          current += ch;
-        }
-      }
-    }
-    parts.push(current.trim());
-    return parts;
-  };
-
-  const headers = parseLine(lines[0]);
-  const rows: Record<string, string>[] = [];
-
-  for (let i = 1; i < lines.length; i++) {
-    const line = lines[i].trim();
-    if (!line) continue;
-    const values = parseLine(line);
-    const row: Record<string, string> = {};
-    headers.forEach((h, idx) => {
-      row[h] = values[idx] ?? '';
-    });
-    rows.push(row);
-  }
-
-  return { headers, rows };
-}
-
-function extractEmails(rows: Record<string, string>[], column: string): string[] {
-  const seen = new Set<string>();
-  const emails: string[] = [];
-  for (const row of rows) {
-    const val = (row[column] ?? '').toLowerCase().trim();
-    if (val && val.includes('@') && val.includes('.') && !seen.has(val)) {
-      seen.add(val);
-      emails.push(val);
-    }
-  }
-  return emails;
-}
-
-function guessEmailColumn(headers: string[], rows: Record<string, string>[]): string | null {
-  const emailHeader = headers.find(h => h.toLowerCase().trim() === 'email');
-  if (emailHeader) return emailHeader;
-
-  let bestCol: string | null = null;
-  let bestCount = 0;
-  for (const h of headers) {
-    let count = 0;
-    for (const row of rows) {
-      const val = (row[h] ?? '').trim();
-      if (val.includes('@') && val.includes('.')) count++;
-    }
-    if (count > bestCount) {
-      bestCount = count;
-      bestCol = h;
-    }
-  }
-  return bestCount > 0 ? bestCol : null;
-}
-
-function parseEmailList(text: string): string[] {
-  const seen = new Set<string>();
-  const emails: string[] = [];
-  // Split on newlines, commas, semicolons, spaces, and tabs
-  const parts = text.split(/[\n,;\s]+/);
-  for (const part of parts) {
-    // Strip surrounding quotes, angle brackets, and whitespace
-    const val = part
-      .replace(/^[<"'\s]+|[>"'\s]+$/g, '')
-      .toLowerCase()
-      .trim();
-    if (val && val.includes('@') && val.includes('.') && !seen.has(val)) {
-      seen.add(val);
-      emails.push(val);
-    }
-  }
-  return emails;
-}
 
 function ineligibleReason(status: string | null): string {
   if (status === null) return 'No subscription - must provision first';
@@ -205,7 +101,7 @@ export function KiloclawExtendTrial() {
   const [pastedText, setPastedText] = useState('');
 
   // Step 1: CSV state
-  const [csvData, setCsvData] = useState<CsvData | null>(null);
+  const [csvData, setCsvData] = useState<CsvTableData | null>(null);
   const [selectedColumn, setSelectedColumn] = useState<string>('');
   const [trialDays, setTrialDays] = useState<string>('7');
   const [isDragging, setIsDragging] = useState(false);
@@ -320,7 +216,7 @@ export function KiloclawExtendTrial() {
   );
 
   const extractedEmails =
-    csvData && selectedColumn ? extractEmails(csvData.rows, selectedColumn) : [];
+    csvData && selectedColumn ? extractEmailsFromColumn(csvData.rows, selectedColumn) : [];
   const csvEmailCount = extractedEmails.length;
 
   // Actions
@@ -376,11 +272,17 @@ export function KiloclawExtendTrial() {
     const content = success
       ? 'email,instance_id,action,new_trial_ends_at\n' +
         filtered
-          .map(r => `${r.email},${r.instanceId ?? ''},${r.action ?? ''},${r.newTrialEndsAt ?? ''}`)
+          .map(r =>
+            [r.email, r.instanceId ?? '', r.action ?? '', r.newTrialEndsAt ?? '']
+              .map(csvField)
+              .join(',')
+          )
           .join('\n')
       : 'email,user_id,instance_id,error\n' +
         filtered
-          .map(r => `${r.email},${r.userId},${r.instanceId ?? ''},${r.error ?? ''}`)
+          .map(r =>
+            [r.email, r.userId, r.instanceId ?? '', r.error ?? ''].map(csvField).join(',')
+          )
           .join('\n');
     downloadCsv(content, `${success ? 'successful' : 'failed'}-trial-extensions.csv`);
   };
@@ -396,9 +298,15 @@ export function KiloclawExtendTrial() {
     const content =
       'email,instance_id,stripe_subscription_id,reason\n' +
       ineligible
-        .map(
-          u =>
-            `${u.email},${u.instanceId ?? ''},${u.stripeSubscriptionId ?? ''},${ineligibleReason(u.subscriptionStatus)}`
+        .map(u =>
+          [
+            u.email,
+            u.instanceId ?? '',
+            u.stripeSubscriptionId ?? '',
+            ineligibleReason(u.subscriptionStatus),
+          ]
+            .map(csvField)
+            .join(',')
         )
         .join('\n');
     downloadCsv(content, 'ineligible-users.csv');
