+++ src/routers/admin/extend-claw-trial-router.ts
@@ -61,6 +61,216 @@ async function fetchLatestSubscriptionPerUser(
   return new Map(rows.map(r => [r.user_id, r]));
 }
 
+type AdminContext = { user: { id: string; google_user_email: string; google_user_name: string | null } };
+
+const EMAIL_TYPES_TO_CLEAR = [
+  'claw_trial_1d',
+  'claw_trial_5d',
+  'claw_suspended_trial',
+  'claw_suspended_subscription',
+  'claw_suspended_payment',
+  'claw_destruction_warning',
+  'claw_instance_destroyed',
+] as const;
+
+async function processOneEmail(
+  email: string,
+  usersByEmail: Map<string, { id: string; email: string }>,
+  latestSubByUserId: Map<string, typeof kiloclaw_subscriptions.$inferSelect>,
+  trialDays: number,
+  ctx: AdminContext
+): Promise<ExtendTrialResult> {
+  const user = usersByEmail.get(email);
+
+  if (!user) {
+    return { email, userId: '', instanceId: null, success: false, error: 'User not found' };
+  }
+
+  const subscription = latestSubByUserId.get(user.id);
+
+  if (!subscription) {
+    return {
+      email,
+      userId: user.id,
+      instanceId: null,
+      success: false,
+      error: 'No KiloClaw subscription found. User must provision an instance first.',
+    };
+  }
+
+  if (subscription.status === 'trialing') {
+    // Extend from the later of current end date or now, so already-expired
+    // trials extend from today rather than from a past date.
+    // Scoped to the specific row id to avoid touching other instances.
+    const [updated] = await db
+      .update(kiloclaw_subscriptions)
+      .set({
+        // Extend from the later of current end date or now, capped at 1 year
+        // from now. The at_limit check at match time prevents users already
+        // past the ceiling from being submitted, but the LEAST here enforces
+        // the ceiling on the resulting value for users currently within it
+        // (e.g. 200 days remaining + 365 days requested = capped at 365).
+        trial_ends_at: sql`LEAST(GREATEST(COALESCE(${kiloclaw_subscriptions.trial_ends_at}::timestamptz, now()), now()) + (${trialDays} * interval '1 day'), now() + interval '1 year')`,
+      })
+      .where(
+        and(
+          eq(kiloclaw_subscriptions.id, subscription.id),
+          eq(kiloclaw_subscriptions.status, 'trialing')
+        )
+      )
+      .returning({ trial_ends_at: kiloclaw_subscriptions.trial_ends_at });
+
+    if (!updated) {
+      // Subscription status changed between match and extend (e.g. user subscribed).
+      return {
+        email,
+        userId: user.id,
+        instanceId: subscription.instance_id,
+        success: false,
+        error: 'Subscription status changed since match — please re-match and retry.',
+      };
+    }
+
+    // Audit log is best-effort — its failure does not undo the extension
+    // and must not cause a false failure result that prompts a retry.
+    try {
+      await createKiloClawAdminAuditLog({
+        action: 'kiloclaw.subscription.bulk_trial_grant',
+        actor_id: ctx.user.id,
+        actor_email: ctx.user.google_user_email,
+        actor_name: ctx.user.google_user_name,
+        target_user_id: user.id,
+        message: `Trial extended by ${trialDays} days via bulk extend, new end: ${updated.trial_ends_at}`,
+        metadata: {
+          source: 'bulk_extend',
+          trialDays,
+          subscriptionId: subscription.id,
+          previousTrialEndsAt: subscription.trial_ends_at,
+          newTrialEndsAt: updated.trial_ends_at,
+          action: 'extended',
+        },
+      });
+    } catch {
+      // Non-fatal
+    }
+
+    return {
+      email,
+      userId: user.id,
+      instanceId: subscription.instance_id,
+      success: true,
+      action: 'extended',
+      newTrialEndsAt: updated.trial_ends_at ?? undefined,
+    };
+  }
+
+  if (subscription.status === 'canceled') {
+    const now = new Date();
+    // Use the same 1-year interval semantics as the SQL ceiling in the trialing path.
+    // addYears is not available without date-fns, so compute via date arithmetic to
+    // match what Postgres's `interval '1 year'` produces (calendar year, not 365 days).
+    const oneYearFromNow = new Date(now);
+    oneYearFromNow.setFullYear(oneYearFromNow.getFullYear() + 1);
+    const newEnd = new Date(Math.min(now.getTime() + trialDays * 86_400_000, oneYearFromNow.getTime()));
+
+    // Resurrect as a fresh trial, mirroring the single-user admin reset path.
+    // Scoped to the specific row id to avoid touching other instances.
+    const [resurrected] = await db
+      .update(kiloclaw_subscriptions)
+      .set({
+        status: 'trialing',
+        plan: 'trial',
+        trial_started_at: now.toISOString(),
+        trial_ends_at: newEnd.toISOString(),
+        stripe_subscription_id: null,
+        stripe_schedule_id: null,
+        scheduled_plan: null,
+        scheduled_by: null,
+        cancel_at_period_end: false,
+        // payment_source and pending_conversion are intentionally left
+        // as-is: stripe_subscription_id = null is the real guard against
+        // stale Stripe webhook replays, and the billing implications of
+        // clearing these fields on a canceled row haven't been fully
+        // analyzed for every payment_source value.
+        current_period_start: null,
+        current_period_end: null,
+        commit_ends_at: null,
+        past_due_since: null,
+        suspended_at: null,
+        destruction_deadline: null,
+      })
+      .where(
+        and(
+          eq(kiloclaw_subscriptions.id, subscription.id),
+          eq(kiloclaw_subscriptions.status, 'canceled')
+        )
+      )
+      .returning({ user_id: kiloclaw_subscriptions.user_id });
+
+    if (!resurrected) {
+      // Subscription status changed between match and extend (e.g. user reactivated).
+      return {
+        email,
+        userId: user.id,
+        instanceId: subscription.instance_id,
+        success: false,
+        error: 'Subscription status changed since match — please re-match and retry.',
+      };
+    }
+
+    // Email log clear and audit log are best-effort — their failure does
+    // not undo the resurrection and must not cause a false failure result
+    // that prompts a retry (which would extend rather than resurrect).
+    try {
+      await db
+        .delete(kiloclaw_email_log)
+        .where(
+          and(
+            eq(kiloclaw_email_log.user_id, user.id),
+            inArray(kiloclaw_email_log.email_type, [...EMAIL_TYPES_TO_CLEAR])
+          )
+        );
+
+      await createKiloClawAdminAuditLog({
+        action: 'kiloclaw.subscription.bulk_trial_grant',
+        actor_id: ctx.user.id,
+        actor_email: ctx.user.google_user_email,
+        actor_name: ctx.user.google_user_name,
+        target_user_id: user.id,
+        message: `Trial restarted for ${trialDays} days via bulk extend (was canceled)`,
+        metadata: {
+          source: 'bulk_extend',
+          trialDays,
+          subscriptionId: subscription.id,
+          previousStatus: subscription.status,
+          newTrialEndsAt: newEnd.toISOString(),
+          action: 'restarted',
+        },
+      });
+    } catch {
+      // Non-fatal
+    }
+
+    return {
+      email,
+      userId: user.id,
+      instanceId: subscription.instance_id,
+      success: true,
+      action: 'restarted',
+      newTrialEndsAt: newEnd.toISOString(),
+    };
+  }
+
+  // Active paid subscription (active, past_due, unpaid) — must not be reset.
+  return {
+    email,
+    userId: user.id,
+    instanceId: subscription.instance_id,
+    success: false,
+    error: `Cannot extend trial: subscription status is "${subscription.status}". Only trialing or canceled subscriptions can be modified.`,
+  };
+}
+
 export const extendClawTrialRouter = createTRPCRouter({
   /**
    * Match a list of emails to existing Kilo user accounts, including their
@@ -107,12 +317,19 @@ export const extendClawTrialRouter = createTRPCRouter({
         if (user) {
           const sub = latestSubByUserId.get(user.id);
           // at_limit when trial already fills the ceiling: extending would be a no-op
-          // because the DB caps at now() + 1 year. Use >= so users exactly at the
-          // boundary are flagged, not just those strictly beyond it.
+          // because the DB caps at now() + 1 year. Compare at day (UTC midnight)
+          // granularity so ms-level clock drift between the Postgres write and this
+          // JS read cannot let an exact-ceiling row slip through as eligible.
+          const todayUtcMidnight = new Date();
+          todayUtcMidnight.setUTCHours(0, 0, 0, 0);
+          const oneYearFromTodayUtc = new Date(todayUtcMidnight);
+          oneYearFromTodayUtc.setUTCFullYear(oneYearFromTodayUtc.getUTCFullYear() + 1);
+          const trialEndDay = sub?.trial_ends_at ? new Date(sub.trial_ends_at) : null;
+          if (trialEndDay) trialEndDay.setUTCHours(0, 0, 0, 0);
           const beyondCeiling =
             sub?.status === 'trialing' &&
-            sub.trial_ends_at !== null &&
-            new Date(sub.trial_ends_at) >= new Date(Date.now() + 365 * 86_400_000);
+            trialEndDay !== null &&
+            trialEndDay >= oneYearFromTodayUtc;
           matched.push({
             email: user.email,
             userId: user.id,
@@ -152,9 +369,8 @@ export const extendClawTrialRouter = createTRPCRouter({
     )
     .mutation(async ({ input, ctx }): Promise<ExtendTrialResult[]> => {
       const { emails, trialDays } = input;
-      const results: ExtendTrialResult[] = [];
 
-      if (emails.length === 0) return results;
+      if (emails.length === 0) return [];
 
       const normalizedEmails = [...new Set(emails.map(e => e.toLowerCase()))];
 
@@ -173,223 +389,21 @@ export const extendClawTrialRouter = createTRPCRouter({
       const userIds = users.map(u => u.id);
       const latestSubByUserId = await fetchLatestSubscriptionPerUser(userIds);
 
-      for (const email of normalizedEmails) {
-        const user = usersByEmail.get(email);
+      // Process all emails concurrently — each email's DB work is independent.
+      const settled = await Promise.allSettled(
+        normalizedEmails.map(email => processOneEmail(email, usersByEmail, latestSubByUserId, trialDays, ctx))
+      );
 
-        if (!user) {
-          results.push({
-            email,
-            userId: '',
-            instanceId: null,
-            success: false,
-            error: 'User not found',
-          });
-          continue;
-        }
-
-        const subscription = latestSubByUserId.get(user.id);
-
-        if (!subscription) {
-          results.push({
-            email,
-            userId: user.id,
-            instanceId: null,
-            success: false,
-            error: 'No KiloClaw subscription found. User must provision an instance first.',
-          });
-          continue;
-        }
-
-        try {
-          if (subscription.status === 'trialing') {
-            // Extend from the later of current end date or now, so already-expired
-            // trials extend from today rather than from a past date.
-            // Scoped to the specific row id to avoid touching other instances.
-            const [updated] = await db
-              .update(kiloclaw_subscriptions)
-              .set({
-                // Extend from the later of current end date or now, capped at 1 year
-                // from now. The at_limit check at match time prevents users already
-                // past the ceiling from being submitted, but the LEAST here enforces
-                // the ceiling on the resulting value for users currently within it
-                // (e.g. 200 days remaining + 365 days requested = capped at 365).
-                trial_ends_at: sql`LEAST(GREATEST(COALESCE(${kiloclaw_subscriptions.trial_ends_at}::timestamptz, now()), now()) + (${trialDays} * interval '1 day'), now() + interval '1 year')`,
-              })
-              .where(
-                and(
-                  eq(kiloclaw_subscriptions.id, subscription.id),
-                  eq(kiloclaw_subscriptions.status, 'trialing')
-                )
-              )
-              .returning({ trial_ends_at: kiloclaw_subscriptions.trial_ends_at });
-
-            if (!updated) {
-              // Subscription status changed between match and extend (e.g. user subscribed).
-              results.push({
-                email,
-                userId: user.id,
-                instanceId: subscription.instance_id,
-                success: false,
-                error: 'Subscription status changed since match — please re-match and retry.',
-              });
-              continue;
-            }
-
-            // Audit log is best-effort — its failure does not undo the extension
-            // and must not cause a false failure result that prompts a retry.
-            try {
-              await createKiloClawAdminAuditLog({
-                action: 'kiloclaw.subscription.bulk_trial_grant',
-                actor_id: ctx.user.id,
-                actor_email: ctx.user.google_user_email,
-                actor_name: ctx.user.google_user_name,
-                target_user_id: user.id,
-                message: `Trial extended by ${trialDays} days via bulk extend, new end: ${updated.trial_ends_at}`,
-                metadata: {
-                  source: 'bulk_extend',
-                  trialDays,
-                  subscriptionId: subscription.id,
-                  previousTrialEndsAt: subscription.trial_ends_at,
-                  newTrialEndsAt: updated.trial_ends_at,
-                  action: 'extended',
-                },
-              });
-            } catch {
-              // Non-fatal
-            }
-
-            results.push({
-              email,
-              userId: user.id,
-              instanceId: subscription.instance_id,
-              success: true,
-              action: 'extended',
-              newTrialEndsAt: updated.trial_ends_at ?? undefined,
-            });
-          } else if (subscription.status === 'canceled') {
-            const now = new Date();
-            const oneYearFromNow = new Date(now.getTime() + 365 * 86_400_000);
-            const newEnd = new Date(
-              Math.min(now.getTime() + trialDays * 86_400_000, oneYearFromNow.getTime())
-            );
-
-            // Resurrect as a fresh trial, mirroring the single-user admin reset path.
-            // Scoped to the specific row id to avoid touching other instances.
-            const [resurrected] = await db
-              .update(kiloclaw_subscriptions)
-              .set({
-                status: 'trialing',
-                plan: 'trial',
-                trial_started_at: now.toISOString(),
-                trial_ends_at: newEnd.toISOString(),
-                stripe_subscription_id: null,
-                stripe_schedule_id: null,
-                scheduled_plan: null,
-                scheduled_by: null,
-                cancel_at_period_end: false,
-                // payment_source and pending_conversion are intentionally left
-                // as-is: stripe_subscription_id = null is the real guard against
-                // stale Stripe webhook replays, and the billing implications of
-                // clearing these fields on a canceled row haven't been fully
-                // analyzed for every payment_source value.
-                current_period_start: null,
-                current_period_end: null,
-                commit_ends_at: null,
-                past_due_since: null,
-                suspended_at: null,
-                destruction_deadline: null,
-              })
-              .where(
-                and(
-                  eq(kiloclaw_subscriptions.id, subscription.id),
-                  eq(kiloclaw_subscriptions.status, 'canceled')
-                )
-              )
-              .returning({ user_id: kiloclaw_subscriptions.user_id });
-
-            if (!resurrected) {
-              // Subscription status changed between match and extend (e.g. user reactivated).
-              results.push({
-                email,
-                userId: user.id,
-                instanceId: subscription.instance_id,
-                success: false,
-                error: 'Subscription status changed since match — please re-match and retry.',
-              });
-              continue;
-            }
-
-            // Email log clear and audit log are best-effort — their failure does
-            // not undo the resurrection and must not cause a false failure result
-            // that prompts a retry (which would extend rather than resurrect).
-            try {
-              const emailTypesToClear = [
-                'claw_trial_1d',
-                'claw_trial_5d',
-                'claw_suspended_trial',
-                'claw_suspended_subscription',
-                'claw_suspended_payment',
-                'claw_destruction_warning',
-                'claw_instance_destroyed',
-              ];
-              await db
-                .delete(kiloclaw_email_log)
-                .where(
-                  and(
-                    eq(kiloclaw_email_log.user_id, user.id),
-                    inArray(kiloclaw_email_log.email_type, emailTypesToClear)
-                  )
-                );
-
-              await createKiloClawAdminAuditLog({
-                action: 'kiloclaw.subscription.bulk_trial_grant',
-                actor_id: ctx.user.id,
-                actor_email: ctx.user.google_user_email,
-                actor_name: ctx.user.google_user_name,
-                target_user_id: user.id,
-                message: `Trial restarted for ${trialDays} days via bulk extend (was canceled)`,
-                metadata: {
-                  source: 'bulk_extend',
-                  trialDays,
-                  subscriptionId: subscription.id,
-                  previousStatus: subscription.status,
-                  newTrialEndsAt: newEnd.toISOString(),
-                  action: 'restarted',
-                },
-              });
-            } catch {
-              // Non-fatal
-            }
-
-            results.push({
-              email,
-              userId: user.id,
-              instanceId: subscription.instance_id,
-              success: true,
-              action: 'restarted',
-              newTrialEndsAt: newEnd.toISOString(),
-            });
-          } else {
-            // Active paid subscription (active, past_due, unpaid) — must not be reset.
-            results.push({
-              email,
-              userId: user.id,
-              instanceId: subscription.instance_id,
-              success: false,
-              error: `Cannot extend trial: subscription status is "${subscription.status}". Only trialing or canceled subscriptions can be modified.`,
-            });
-          }
-        } catch (error) {
-          results.push({
-            email,
-            userId: user.id,
-            instanceId: subscription.instance_id,
-            success: false,
-            error: error instanceof Error ? error.message : 'Unknown error',
-          });
-        }
-      }
-
-      return results;
+      return settled.map((outcome, i) => {
+        if (outcome.status === 'fulfilled') return outcome.value;
+        const email = normalizedEmails[i];
+        return {
+          email,
+          userId: usersByEmail.get(email)?.id ?? '',
+          instanceId: null,
+          success: false,
+          error: outcome.reason instanceof Error ? outcome.reason.message : 'Unknown error',
+        };
+      });
     }),
 });
